modified:   ../api_static/plasmapy.dispersion.dispersionfunction.rst 	deleted:    ../api_static/plasmapy.dispersion.rst 	new file:   ../api_static/plasmapy.dispersion.two_fluid_dispersion.rst 	deleted:    dispersionfunction.rst 	modified:   index.rst 	deleted:    twofluiddispersionsolver.rst 	renamed:    ../notebooks/dispersion/two_fluid_dispersion_solver.ipynb -> ../notebooks/dispersion/two_fluid_dispersion.ipynb 	modified:   ../../plasmapy/dispersion/__init__.py 	renamed:    ../../plasmapy/dispersion/tests/test_two_fluid_dispersion_solver.py -> ../../plasmapy/dispersion/tests/test_two_fluid_dispersion.py 	renamed:    ../../plasmapy/dispersion/two_fluid_dispersion_solver.py -> ../../plasmapy/dispersion/two_fluid_dispersion.py
