--- docs/api_static/plasmapy.dispersion.rst
@@ -1,29 +1,0 @@
-:orphan:
-
-`plasmapy.dispersion`
-=====================
-
-.. currentmodule:: plasmapy.dispersion
-
-.. automodule:: plasmapy.dispersion
-
-Sub-Packages & Modules
-----------------------
-
-.. autosummary::
-
-   dispersionfunction
-   two_fluid_dispersion_solver
-
-.. automodapi:: plasmapy.dispersion
-   :no-main-docstr:
-   :no-heading:
-   :include-all-objects:
-
-Example Notebooks
------------------
-
-.. nbgallery::
-
-    /notebooks/dispersion/dispersion_function
-	/notebooks/dispersion/two_fluid_dispersion_solver
