+++ docs/api_static/plasmapy.dispersion.two_fluid_dispersion.rst
@@ -1,0 +1,18 @@
+:orphan:
+
+`plasmapy.dispersion.two_fluid_dispersion`
+==========================================
+
+.. currentmodule:: plasmapy.dispersion.two_fluid_dispersion
+
+.. automodapi:: plasmapy.dispersion.two_fluid_dispersion
+   :include-all-objects:
+   :no-heading:
+   :headings: "-^"
+
+Example Notebooks
+-----------------
+
+.. nbgallery::
+
+    /notebooks/dispersion/two_fluid_dispersion
