--- docs/dispersion/dispersionfunction.rst
@@ -1,15 +1,0 @@
-.. _dispersionfunction:
-
-***************************************************************
-Dispersion relations (`plasmapy.dispersion.dispersionfunction`)
-***************************************************************
-
-.. currentmodule:: plasmapy.dispersion.dispersionfunction
-
-.. automodapi:: plasmapy.dispersion.dispersionfunction
-   :no-heading:
-
-.. nbgallery::
-    :caption: Examples
-
-    /notebooks/dispersion/dispersion_function
