--- docs/dispersion/twofluiddispersionsolver.rst
@@ -1,15 +1,0 @@
-.. _two_fluid_dispersion_solver:
-
-*********************************************************************
-Dispersion solver (`plasmapy.dispersion.two_fluid_dispersion_solver`)
-*********************************************************************
-
-.. currentmodule:: plasmapy.dispersion.two_fluid_dispersion_solver
-
-.. automodapi:: plasmapy.dispersion.two_fluid_dispersion_solver
-   :no-heading:
-
-.. nbgallery::
-    :caption: Examples
-
-    /notebooks/dispersion/two_fluid_dispersion_solver
